Update windows config to add backgroundtasks
No point having system media controls when the app cannot play any audio
in the background or when minimized.

Add extensions for background audio tasks.

Must also register media controls (which this plugin should do) and use
a version of the media plugin that creates the media element with the
correct msAudioCategory (which my repo has and has been pushed to master
as a PR)